Check for errors in return hash string
* Fix check for errors

Some implementations of crypt (e.g. libxcrypt as shipped in at least
Fedora and openSUSE) instead of returning NULL on errors might also
return an invalid hash string starting with '*' for indicating an error.
This add some checks for that error condition.

Fixes #8

* Fix builds on non-glibc Linux

Allow to build on non-glibc based Linux systems (as e.g. Alpine). Make
the glibc version check a private function as this is only supposed to
be used by the tests. Also make the test more explicit and remove the
broken lexical comparison on the minor version.

Fixes #10

* Explicitly error out on windows and darwin

Obviously this module does not work on Windows. Also it seems the the
Darwin variant of crypt is different enough from the BSD and Linux
variants to be easily supportable.

This adds a dummy implementation that just errors out on those
platforms.